# Reduce allocations and expensive operations parsing yaml