--- a/src/main/java/org/yaml/snakeyaml/Yaml.java
+++ b/src/main/java/org/yaml/snakeyaml/Yaml.java
@@ -86,6 +86,16 @@ public class Yaml {
   /**
    * Create Yaml instance.
    *
+   * @param loadingConfig LoadingConfig to control load behavior
+   * @param dumperOptions DumperOptions to configure outgoing objects
+   */
+  public Yaml(LoaderOptions loadingConfig, DumperOptions dumperOptions) {
+    this(new Constructor(loadingConfig), new Representer(dumperOptions), dumperOptions);
+  }
+
+  /**
+   * Create Yaml instance.
+   *
    * @param representer Representer to emit outgoing objects. The DumperOptions may not be fully
    *        respected. It is better to use explicit DumperOptions.
    * @deprecated use method with explicit DumperOptions