--- a/src/main/java/org/yaml/snakeyaml/composer/Composer.java
+++ b/src/main/java/org/yaml/snakeyaml/composer/Composer.java
@@ -20,6 +20,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+
 import org.yaml.snakeyaml.DumperOptions.FlowStyle;
 import org.yaml.snakeyaml.LoaderOptions;
 import org.yaml.snakeyaml.comments.CommentEventsCollector;
@@ -42,6 +43,7 @@ import org.yaml.snakeyaml.nodes.SequenceNode;
 import org.yaml.snakeyaml.nodes.Tag;
 import org.yaml.snakeyaml.parser.Parser;
 import org.yaml.snakeyaml.resolver.Resolver;
+import org.yaml.snakeyaml.util.Tuple;
 
 /**
  * Creates a node graph from parser events.
@@ -315,12 +317,14 @@ public class Composer {
       node.setAnchor(anchor);
       anchors.put(anchor, node);
     }
+    boolean hasMerges = false;
     while (!parser.checkEvent(Event.ID.MappingEnd)) {
       blockCommentsCollector.collectEvents();
       if (parser.checkEvent(Event.ID.MappingEnd)) {
         break;
       }
       composeMappingChildren(children, node);
+      hasMerges = hasMerges || node.isMerged();
     }
     if (startEvent.isFlow()) {
       node.setInLineComments(inlineCommentsCollector.collectEvents().consume());
@@ -331,6 +335,12 @@ public class Composer {
     if (!inlineCommentsCollector.isEmpty()) {
       node.setInLineComments(inlineCommentsCollector.consume());
     }
+
+    if (loadingConfig.isMergeOnCompose() && hasMerges) {
+      List<NodeTuple> updatedValue = flatten(node);
+      node.setValue(updatedValue);
+    }
+
     return node;
   }
 
@@ -347,6 +357,80 @@ public class Composer {
     }
     Node itemValue = composeValueNode(node);
     children.add(new NodeTuple(itemKey, itemValue));
+  }
+
+  protected List<NodeTuple> flatten(MappingNode node) {
+    List<NodeTuple> original = node.getValue();
+    Set<String> keys = new HashSet<>(original.size());
+    List<NodeTuple> updated = new ArrayList<>(original.size());
+    List<NodeTuple> merges = new ArrayList<>(2);
+
+    for (NodeTuple tuple : original) {
+      Node keyNode = tuple.getKeyNode();
+      if (keyNode.getTag().equals(Tag.MERGE)) {
+        merges.add(tuple);
+      } else {
+        updated.add(tuple);
+        if (keyNode instanceof ScalarNode) {
+          ScalarNode sNode = (ScalarNode) keyNode;
+          keys.add(sNode.getValue());
+        }
+      }
+    }
+
+    for (NodeTuple tuple : merges) {
+      Node valueNode = tuple.getValueNode();
+      if (valueNode instanceof SequenceNode) {
+        SequenceNode seqNode = (SequenceNode) valueNode;
+        for (Node ref : seqNode.getValue()) {
+          MappingNode mergable = asMappingNode(ref);
+          Tuple<List<NodeTuple>, Set<String>> filtered = filter(mergable.getValue(), keys);
+          updated.addAll(filtered._1());
+          keys.addAll(filtered._2());
+        }
+      } else {
+        MappingNode mergable = asMappingNode(valueNode);
+        Tuple<List<NodeTuple>, Set<String>> filtered = filter(mergable.getValue(), keys);
+        updated.addAll(filtered._1());
+        keys.addAll(filtered._2());
+      }
+    }
+    return updated;
+  }
+
+  protected MappingNode asMappingNode(Node node) {
+    if (node instanceof MappingNode) {
+      return (MappingNode) node;
+    } else {
+      Node ref = anchors.get(node.getAnchor());
+      if (ref instanceof MappingNode) {
+        return (MappingNode) ref;
+      }
+    }
+    // TODO: add corrrect marks to the exception
+    throw new ComposerException("Expected mapping node or an anchor referencing mapping", null,
+        null, null);
+  }
+
+  protected Tuple<List<NodeTuple>, Set<String>> filter(List<NodeTuple> mergables,
+      Set<String> filter) {
+    int size = mergables.size();
+    Set<String> keys = new HashSet<>(size);
+    List<NodeTuple> result = new ArrayList<>(size);
+    for (NodeTuple tuple : mergables) {
+      Node key = tuple.getKeyNode();
+      if (key instanceof ScalarNode) {
+        ScalarNode sNode = (ScalarNode) key;
+        String nodeValue = sNode.getValue();
+        if (!filter.contains(nodeValue)) {
+          result.add(tuple);
+          keys.add(nodeValue);
+        }
+      } else {
+        result.add(tuple);
+      }
+    }
+    return new Tuple<>(result, keys);
   }
 
   /**