--- a/src/main/java/org/yaml/snakeyaml/constructor/BaseConstructor.java
+++ b/src/main/java/org/yaml/snakeyaml/constructor/BaseConstructor.java
@@ -44,6 +44,7 @@ import org.yaml.snakeyaml.nodes.NodeTuple;
 import org.yaml.snakeyaml.nodes.ScalarNode;
 import org.yaml.snakeyaml.nodes.SequenceNode;
 import org.yaml.snakeyaml.nodes.Tag;
+import org.yaml.snakeyaml.util.Tuple;
 
 /**
  * Base code
@@ -80,8 +81,8 @@ public abstract class BaseConstructor {
   protected Composer composer;
   final Map<Node, Object> constructedObjects;
   private final Set<Node> recursiveObjects;
-  private final ArrayList<RecursiveTuple<Map<Object, Object>, RecursiveTuple<Object, Object>>> maps2fill;
-  private final ArrayList<RecursiveTuple<Set<Object>, Object>> sets2fill;
+  private final ArrayList<Tuple<Map<Object, Object>, Tuple<Object, Object>>> maps2fill;
+  private final ArrayList<Tuple<Set<Object>, Object>> sets2fill;
 
   /**
    * the tag for the root node
@@ -119,9 +120,8 @@ public abstract class BaseConstructor {
     }
     constructedObjects = new HashMap<Node, Object>();
     recursiveObjects = new HashSet<Node>();
-    maps2fill =
-        new ArrayList<RecursiveTuple<Map<Object, Object>, RecursiveTuple<Object, Object>>>();
-    sets2fill = new ArrayList<RecursiveTuple<Set<Object>, Object>>();
+    maps2fill = new ArrayList<Tuple<Map<Object, Object>, Tuple<Object, Object>>>();
+    sets2fill = new ArrayList<Tuple<Set<Object>, Object>>();
     typeDefinitions = new HashMap<Class<? extends Object>, TypeDescription>();
     typeTags = new HashMap<Tag, Class<? extends Object>>();
 
@@ -219,14 +219,14 @@ public abstract class BaseConstructor {
    */
   private void fillRecursive() {
     if (!maps2fill.isEmpty()) {
-      for (RecursiveTuple<Map<Object, Object>, RecursiveTuple<Object, Object>> entry : maps2fill) {
-        RecursiveTuple<Object, Object> key_value = entry._2();
+      for (Tuple<Map<Object, Object>, Tuple<Object, Object>> entry : maps2fill) {
+        Tuple<Object, Object> key_value = entry._2();
         entry._1().put(key_value._1(), key_value._2());
       }
       maps2fill.clear();
     }
     if (!sets2fill.isEmpty()) {
-      for (RecursiveTuple<Set<Object>, Object> value : sets2fill) {
+      for (Tuple<Set<Object>, Object> value : sets2fill) {
         value._1().add(value._2());
       }
       sets2fill.clear();
@@ -593,7 +593,7 @@ public abstract class BaseConstructor {
    * not observe key hashCode changes.
    */
   protected void postponeMapFilling(Map<Object, Object> mapping, Object key, Object value) {
-    maps2fill.add(0, new RecursiveTuple(mapping, new RecursiveTuple(key, value)));
+    maps2fill.add(0, new Tuple(mapping, new Tuple(key, value)));
   }
 
   protected void constructSet2ndStep(MappingNode node, Set<Object> set) {
@@ -623,7 +623,7 @@ public abstract class BaseConstructor {
    * does not observe value hashCode changes.
    */
   protected void postponeSetFilling(Set<Object> set, Object key) {
-    sets2fill.add(0, new RecursiveTuple<Set<Object>, Object>(set, key));
+    sets2fill.add(0, new Tuple<Set<Object>, Object>(set, key));
   }
 
   public void setPropertyUtils(PropertyUtils propertyUtils) {
@@ -658,25 +658,6 @@ public abstract class BaseConstructor {
     typeTags.put(tag, definition.getType());
     definition.setPropertyUtils(getPropertyUtils());
     return typeDefinitions.put(definition.getType(), definition);
-  }
-
-  private static class RecursiveTuple<T, K> {
-
-    private final T _1;
-    private final K _2;
-
-    public RecursiveTuple(T _1, K _2) {
-      this._1 = _1;
-      this._2 = _2;
-    }
-
-    public K _2() {
-      return _2;
-    }
-
-    public T _1() {
-      return _1;
-    }
   }
 
   public final boolean isExplicitPropertyUtils() {