… src/changes/changes.xml
… src/main/java/org/yaml/snakeyaml/util/Tuple.java
… src/test/java/org/yaml/snakeyaml/issues/issue1096/MergeExpandTest.java
… src/test/java/org/yaml/snakeyaml/issues/issue1096/MergeOnComposeTest.java
… src/test/java/org/yaml/snakeyaml/issues/issue1096/MergeOnSerializeTest.java
… src/test/java/org/yaml/snakeyaml/jmh/ParseBenchmark.java
… src/test/resources/issues/issue1096-complex-merge-input.yaml
… src/test/resources/issues/issue1096-complex-merge-output.yaml
… src/test/resources/issues/issue1096-merge-input.yaml
… src/test/resources/issues/issue1096-merge-output.yaml
… src/test/resources/issues/issue1096-simple-merge-input.yaml
… src/test/resources/issues/issue1096-simple-merge-output.yaml