@schlosna 2024-08-15
moved to https://github.com/snakeyaml/snakeyaml/pull/15